New Tab Page: use DataSource->Add{String, Integer, Boolean} instead of rvh->SetWebUIProperty
wvh->SetWebUIProperty has issues in that the rvh does not always fire the 'ready' event, especially when navigating 'back'. Using DataSource is a more consistent chromium method for sending properties to the JS context. However, this is usually used for non-changing data. Whilst an UpdateDataSource function does exist, that kind of data is usually pulled from the JS via an explicit call to the C++ WebUI, so we may have to go down that route. The issue with the implementation in this commit is that a page may miss the 'updated' events that fire, for example due to have being navigated away and then back to the page.

So whilst this fix does ensure that the page receives some data, it does not ensure that it gets up to date data. That should be fixed via calling DataSource::Update, or by using explicit data retrieval JS -> C++ functions.